[feat] move compare_results to utils
