+++ scalabel/label/utils.py
@@ -3,12 +3,13 @@ import math
 from typing import Dict, List
 
 import numpy as np
+import pytest
 from scipy.spatial.transform import Rotation
 
 from scalabel.common.typing import NDArrayF64
 
 from ..common.typing import NDArrayF64
-from .typing import Category, Extrinsics, Intrinsics, Label
+from .typing import Category, Extrinsics, Frame, Intrinsics, Label
 
 
 def get_intrinsics_from_matrix(matrix: NDArrayF64) -> Intrinsics:
@@ -123,3 +124,64 @@ def rotation_y_to_alpha(
     if alpha <= -math.pi:
         alpha += 2 * math.pi
     return alpha
+
+
+def compare_results(result: List[Frame], result_compare: List[Frame]) -> None:
+    """Compare two list of frames."""
+    for frame, frame_ref in zip(result, result_compare):
+        assert frame.name == frame_ref.name
+        assert frame.video_name == frame_ref.video_name
+        assert frame.frame_index == frame_ref.frame_index
+        if frame.intrinsics is not None:
+            assert frame_ref.intrinsics is not None
+            assert frame.intrinsics.focal == pytest.approx(
+                frame_ref.intrinsics.focal
+            )
+            assert frame.intrinsics.center == pytest.approx(
+                frame_ref.intrinsics.center
+            )
+            assert frame.intrinsics.skew == pytest.approx(
+                frame_ref.intrinsics.skew
+            )
+        else:
+            assert frame.intrinsics == frame_ref.intrinsics
+        if frame.extrinsics is not None:
+            assert frame_ref.extrinsics is not None
+            assert frame.extrinsics.location == pytest.approx(
+                frame_ref.extrinsics.location
+            )
+            assert frame.extrinsics.rotation == pytest.approx(
+                frame_ref.extrinsics.rotation
+            )
+        else:
+            assert frame.extrinsics == frame_ref.extrinsics
+
+        if frame.labels is not None:
+            assert frame_ref.labels is not None
+            for label, label_ref in zip(frame.labels, frame_ref.labels):
+                assert label.id == label_ref.id
+                assert label.category == label_ref.category
+                if label.box2d is not None:
+                    assert label_ref.box2d is not None
+                    assert label.box2d.x1 == pytest.approx(label_ref.box2d.x1)
+                    assert label.box2d.y1 == pytest.approx(label_ref.box2d.y1)
+                    assert label.box2d.x2 == pytest.approx(label_ref.box2d.x2)
+                    assert label.box2d.y2 == pytest.approx(label_ref.box2d.y2)
+                else:
+                    assert label.box2d == label_ref.box2d
+
+                if label.box3d is not None:
+                    assert label_ref.box3d is not None
+                    assert label.box3d.location == pytest.approx(
+                        label_ref.box3d.location
+                    )
+                    assert label.box3d.dimension == pytest.approx(
+                        label_ref.box3d.dimension
+                    )
+                    assert label.box3d.orientation == pytest.approx(
+                        label_ref.box3d.orientation
+                    )
+                else:
+                    assert label.box3d == label_ref.box3d
+        else:
+            assert frame.labels == frame_ref.labels
